backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rule-engine"
version = "0.1.0"
description = "Real-time bounding box processing pipeline engine"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "Pillow>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "isort>=5.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["rule_execution_engine*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "RET", "SIM"]
ignore = ["E501"]
target-version = "py39"

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=rule_execution_engine --cov-report=term-missing --cov-fail-under=80"
